+++ src/lib/Warp.NET.SourceGen/Generators/GameGenerator.cs
@@ -27,15 +27,39 @@ public class GameGenerator : IIncrementalGenerator
 
 		namespace {{_namespace}};
 
-		public partial class Game
+		public partial class Game : GameBase, IGameBase<Game>
 		{
+			private static Game? _self;
+
 			{{_gameObjectListFields}}
+
+			private Game(string initialWindowTitle, int initialWindowWidth, int initialWindowHeight, bool initialWindowFullScreen)
+				: base(initialWindowTitle, initialWindowWidth, initialWindowHeight, initialWindowFullScreen)
+			{
+			}
+
+			public static Game Self
+			{
+				get => _self ?? throw new InvalidOperationException("Game is not initialized.");
+				set
+				{
+					if (_self != null)
+						throw new InvalidOperationException("Game is already initialized.");
+
+					_self = value;
+				}
+			}
 
 			{{_gameObjectListProperties}}
 
 			{{_menuProperties}}
 
 			{{_singletonProperties}}
+
+			public static Game Construct(string initialWindowTitle, int initialWindowWidth, int initialWindowHeight, bool initialWindowFullScreen)
+			{
+				return new(initialWindowTitle, initialWindowWidth, initialWindowHeight, initialWindowFullScreen);
+			}
 
 			protected override void HandleAdds({{Constants.RootNamespace}}.GameObjects.IGameObject gameObject)
 			{